Moved chat stuff to chat fragment
