+++ app/src/main/java/com/denweisenseel/scotlandsaarexperimental/ChatFragment.java
@@ -2,9 +2,13 @@ package com.denweisenseel.scotlandsaarexperimental;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v4.content.ContextCompat;
+import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
@@ -20,6 +24,7 @@ import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
+import com.aurelhubert.ahbottomnavigation.notification.AHNotification;
 import com.denweisenseel.scotlandsaarexperimental.adapter.ChatMessageAdapter;
 import com.denweisenseel.scotlandsaarexperimental.api.RequestBuilder;
 import com.denweisenseel.scotlandsaarexperimental.data.ChatDataParcelable;
@@ -28,7 +33,9 @@ import com.google.firebase.iid.FirebaseInstanceId;
 
 import org.json.JSONObject;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 
 
 public class ChatFragment extends Fragment {
@@ -37,12 +44,20 @@ public class ChatFragment extends Fragment {
     private ChatMessageAdapter cAdapater;
     private final ArrayList<ChatDataParcelable> chatList =  new ArrayList();
 
+    private BroadcastReceiver chatMessageReceiver;
+
     private final String TAG = "GameChat";
 
     String gameId;
 
     public ChatFragment() {
         // Required empty public constructor
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_chat, container, false);
     }
 
     public static ChatFragment newInstance(String gameId) {
@@ -60,26 +75,18 @@ public class ChatFragment extends Fragment {
             this.gameId = getArguments().getString("gameId");
         }
 
-
     }
 
     @Override
     public void onViewCreated(View v, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(v, savedInstanceState);
-
-
+        Log.i(TAG, "Chat created");
         ListView listView = v.findViewById(R.id.chat_fragment_listview);
-
         cAdapater = new ChatMessageAdapter(getActivity(), chatList);
-
-
         cAdapater.notifyDataSetChanged();
         listView.setAdapter(cAdapater);
-
         listView.setTranscriptMode(AbsListView.TRANSCRIPT_MODE_ALWAYS_SCROLL);
         listView.setStackFromBottom(true);
-
-
         final EditText chatMessageInput = v.findViewById(R.id.chat_fragment_chatMessage);
 
         chatMessageInput.setOnEditorActionListener(new EditText.OnEditorActionListener() {
@@ -91,8 +98,6 @@ public class ChatFragment extends Fragment {
                     sendChatMessage(chatMessageInput.getText().toString());
                     chatMessageInput.setText("");
                     chatMessageInput.findFocus();
-
-
                     return true;
                 }
                 return false;
@@ -100,21 +105,13 @@ public class ChatFragment extends Fragment {
         });
 
 
-
+        initChatPushReceivers();
     }
 
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-
-        return inflater.inflate(R.layout.fragment_chat, container, false);
-    }
 
     @Override
     public void onStart() {
         super.onStart();
-        mListener.retrieveChatMessages();
     }
 
     @Override
@@ -140,17 +137,13 @@ public class ChatFragment extends Fragment {
     }
 
     public interface ChatFragmentInteractionListener {
-
-        void retrieveChatMessages();
+        void onMessageReceived();
     }
-
 
     private void sendChatMessage(final String message) {
 
         String firebaseToken = FirebaseInstanceId.getInstance().getToken();
         String[] args = {gameId,firebaseToken, message};
-
-
         JsonObjectRequest gameRequest = new JsonObjectRequest(Request.Method.POST, RequestBuilder.buildRequestUrl(RequestBuilder.CHAT_MESSAGE, args ),null, new Response.Listener<JSONObject>() {
             @Override
             public void onResponse(JSONObject response) {
@@ -162,8 +155,29 @@ public class ChatFragment extends Fragment {
                 Log.e(TAG, error.toString());
             }
         });
-
         VolleyRequestQueue.getInstance(getActivity()).addToRequestQueue(gameRequest);
+    }
 
+
+    private void initChatPushReceivers() {
+        chatMessageReceiver = new BroadcastReceiver() {
+            @Override
+            public void onReceive(Context context, Intent intent) {
+                if (intent.getAction().equals(getString(R.string.LOBBY_PLAYER_JOIN))) {
+                    ArrayList<String> argList = intent.getStringArrayListExtra(getString(R.string.BROADCAST_DATA));
+                    String playerName = argList.get(0);
+                    ChatDataParcelable chatMessage = new ChatDataParcelable(playerName, "joined the lobby", new SimpleDateFormat("HH.mm").format(new Date()));
+                    sendMessage(chatMessage);
+                } else if (intent.getAction().equals(getString(R.string.LOBBY_PLAYER_MESSAGE))) {
+                    ArrayList<String> argList = intent.getStringArrayListExtra(getString(R.string.BROADCAST_DATA));
+                    ChatDataParcelable chatMessage = new ChatDataParcelable(argList.get(0), argList.get(1), argList.get(2));
+                    sendMessage(chatMessage);
+                    mListener.onMessageReceived();
+                }
+            }
+        };
+
+        LocalBroadcastManager.getInstance(getActivity()).registerReceiver(chatMessageReceiver, new IntentFilter(getString(R.string.LOBBY_PLAYER_JOIN)));
+        LocalBroadcastManager.getInstance(getActivity()).registerReceiver(chatMessageReceiver, new IntentFilter(getString(R.string.LOBBY_PLAYER_MESSAGE)));
     }
 }
